fix: raise ValueError when convert from pandas when index name duplicates column name (#15938) - handle multiindex
